+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     * Fixed empty `route()` and `raw_route()` when getting translated pages [#3184](https://github.com/getgrav/grav/pull/3184)
     * Fixed error on `bin/gpm plugin uninstall` [#3207](https://github.com/getgrav/grav/issues/3207)
     * Fixed lowering uppercase characters in usernames when saving from frontend [#2565](https://github.com/getgrav/grav/pull/2565)
+    * Fixed broken min/max validation for field `type: int`
 
 # v1.7.5
 ## 02/01/2021